Avoid NullReferenceException when cancelling disposed MySqlCommand
Currently if MySqlCommand was asynchronously executed with a CancellationToken, and that token is cancelled After the command is disposed, the MySqlCommand throws NullReferenceException because at that point command's connection is null. This is OK for release builds because the exception would be caught by System.Data.Common.DbCommand's internal logic, but it is very annoying to see these exceptions when debugging. Additionally, it ads extra unnecessary overhead of throwing/catching an exception. The fix is simply to check whether connection is null before calling CancelQuery.